babeld: new package
Babel is a loop-avoiding distance-vector routing protocol for IPv6 and
IPv4 with fast convergence properties. It is based on the ideas in DSDV,
AODV and Cisco's EIGRP, but is designed to work well not only in wired
networks but also in wireless mesh networks, and has been extended with
support for overlay networks.

https://www.irif.univ-paris-diderot.fr/~jch/software/babel/

Signed-off-by: Christian Stewart <christian@paral.in>