docker-engine: fix handling of vendor tree
The latest docker-engine release changes the vendor tree slightly. The Go tool
will recognize a ./vendor directory and use it as a vendor tree
automatically, but only when run inside a valid GOPATH. This patch
adjusts how the GOPATH is built - now docker/docker is linked into a
blank tree, and the Go tool recognizes the ./vendor directory
correctly.

The deleted patch is no longer necessary as of the last release and
causes a build failure on my machine.

Signed-off-by: Christian Stewart <christian@paral.in>