docker-engine: bump to v17.07.0-ce
This commit bumps docker-engine to v17.07.0-ce.

The docker-engine repository has now been moved to:

  https://github.com/docker/docker-ce

The CLI and the daemon have been split into separate repositories:

  - https://github.com/docker/cli
  - https://github.com/moby/moby

The daemon is still in the "moby" repo, but the import paths in Go are:

  github.com/docker/docker

As before. Presumably, the maintainers will eventually split the engine out
to a different repository as well.

Please note that they literally copy the entire docker/cli and moby/moby
repositories into the docker-ce repository, as well as all of their
dependencies in the vendor directories. This means that downloading the
docker-ce source tarball contains:

 - The cli sources
 - The daemon sources
 - Two copies of all of the dependencies of both

At the moment this is unavoidable, but does produce a working build.

The patch was removed, as it is included in this release.

Signed-off-by: Christian Stewart <christian@paral.in>