docker-engine: don't link against libsystemd without systemd
Docker links against libsystemd unconditionally in the latest release. This is
entirely unnecessary, and the included patch disables this behavior in the
absence of the systemd build tag, which is now set when systemd is enabled in
buildroot.

This fixes linking errors when building against sysv, and avoids adding a
hard dependency on systemd in Docker.

Docker-Engine should also use the target pkgconfig, not the host pkgconfig. This
is fixed in this patch.

Signed-off-by: Christian Stewart <christian@paral.in>