package/docker-cli: re-introduce static option
The option to compile the docker-cli binary statically (for use with
bind-mounting the CLI binary into containers for accessing the parent Docker
API) was previously removed due to an incompatibility with static compilation
and cgo.

This commit re-introduces the option by adding the osusergo, netgo, and
CGO_ENABLED=0 tags and variables, which disable cgo implementations of user
lookup and network lookups respectively.

Signed-off-by: Christian Stewart <christian@paral.in>