package/docker-cli: re-introduce static option
The option to compile the docker-cli binary statically (for use with
bind-mounting the CLI binary into containers for accessing the parent Docker
API) was previously removed due to an incompatibility with static compilation
and cgo.

This commit re-introduces the option, disabling the cgo calls that were
previously problematic with the osusergo tag.

Signed-off-by: Christian Stewart <christian@paral.in>