runc: bump to version 1.0.0-rc6
Previously, a specific commit hash from the Docker runc.installer was used to
determine the required runc version for the Docker Engine. This old commit hash
used was a untagged pre-1.0.0 release of runc, closer to an earlier release
candidate.

The runc version used in the Debian distribution is not the pinned version
previously used by Buildroot. It is the latest release candidate. The latest
release candidate is known to be compatible with the Docker Engine, and there is
no justification for pinning to an older RC anymore.

This commit bumps to the latest RC, 1.0.0-rc6. A v1.0.0 is expected soon.

Signed-off-by: Christian Stewart <christian@paral.in>