package/go: switch to go modules from gopath
This commit moves to the new (as of go 1.11) modules system in Go for managing
resolving import paths. Previously, Buildroot created a fake GOPATH tree and
symbolic-linked the source code into this tree. With this patch, the Go compiler
recognizes the import path of the root package correctly with the presence of a
go.mod file, and the GOPATH setup steps are no longer required.

Existing packages will compile without changes as the vendor/ tree is still the
default approach for including package dependencies. Output of the build process
will be the same as with GOPATH.

Some Go packages contain $(2)_WORKSPACE variables. These are no longer required.

Signed-off-by: Christian Stewart <christian@paral.in>