docker-cli: upgrade to go modules
This commit introduces go.mod and go.sum files derived from the vendor.conf in
the upstream Docker cli repository.

Replace directives are added to indicate the correct version of docker/engine to
pull in. Additionally, a patch is added to bring in the upstream modifications
to the vendored spf13/pflag package.

Signed-off-by: Christian Stewart <christian@paral.in>