python: use system libffi for host-python
python2 contains a bundled copy of libffi which is currently out of
sync with the latest libffi release.
There is an option to use a system libffi, buildroot already uses it for
the target python2 build and for python3. In python3, the bundled copy
doesn't exist anymore and the system-provided libffi is required.

The bundled copy currently fails to build on aarch64 host due to a
missing definition of AARCH64_CALL_CONTEXT_SIZE. This define was removed
from the headers in recent libffi releases and the host compiler might
be including the system headers before the bundled headers.

To solve this and since buildroot already relies on system libffi for
target python2 and python3 anyway, switch host python2 to use system
libffi.

Signed-off-by: Erico Nunes <nunes.erico@gmail.com>