package/rtl8821au: switch to aircrack-ng fork
Switching to the aircrack-ng fork of rtl8821au USB WiFi kernel module.

This fixes build errors against 5.11.

This package has utilized several forks over the years, the aircrack-ng version
is currently the most well maintained, and most recently updated. This patch
also drops all of the patches that were previously maintained in the Buildroot
tree, which are no longer necessary.

Signed-off-by: Christian Stewart <christian@paral.in>