package/rtl8821au: switch to lwfinger fork
Switching to the lwfinger fork of rtl8821au USB WiFi kernel module.

This fixes build errors against kernel 5.18.

This patch also drops all of the patches that were previously maintained in the
Buildroot tree, which are no longer necessary.

It should be noted that rtl8812au-aircrack-ng provides the same driver.

Signed-off-by: Christian Stewart <christian@paral.in>