package/pkg-golang: host-go is also a build dependency
Currently, host-go is a download dependency, because we need go for the
vendoring during the download step.

However, when using a package in override-srcdir, there is no download
step, so host-go is not pulled in as a dependency, which will break the
build of a single package from scratch.

We fix that by adding host-go as a standard dependency too.

Signed-off-by: Yann E. MORIN <yann.morin@orange.com>
Cc: Thomas Petazzoni <thomas.petazzoni@bootlin.com>
Cc: Christian Stewart <christian@paral.in>
Cc: Anisse Astier <anisse@astier.eu>
Acked-by: Thomas Petazzoni <thomas.petazzoni@bootlin.com>
Reviewed-by: Christian Stewart <christian@paral.in>