package/pkg-golang: fix build with per-package directories
Build with per-package directory is broken, because go is not found in
the PATH, when trying to download and vendor a go package.

This is because FOO_DL_ENV contains $(HOST_GO_COMMON_ENV), which defines
the PATH as PATH=$(BR_PATH). This is correct, except this is expanded at
the time the golang-package macro is evaluated, which means PATH
contains the 'global' BR_PATH, i.e.: $(O)/host/bin:$(O)/host/sbin:...

However, with PPD, this does not yet exist at build time; only the
per-package hoqt directory exists.

We want to have it expanded at the time the recipe is run, so like all
other variables, we need to $$-expand it.

At the same time, also $$-expand two other variables (even though those
are benign, it is better for consistency that they be $$-expanded).

Signed-off-by: Yann E. MORIN <yann.morin@orange.com>
Cc: Thomas Petazzoni <thomas.petazzoni@bootlin.com>
Cc: Anisse Astier <anisse@astier.eu>
Cc: Christian Stewart <christian@paral.in>