package/pkg-utils: prevent kconfig_enable_opt from changing =m to =y
The KCONFIG_ENABLE_OPT is intended to enable a required kernel configuration
option when a package requires it.

However, this will often override an existing enabled module with `=m` with `=y`
which overrides the module to be built-in instead of separate.

This is undesirable behavior; we often want these as `=m` and not `=y` to reduce
the size of the kernel image.

This patch changes KCONFIG_MUNGE_DOT_CONFIG to prevent changing `=m` to `=y`.

Signed-off-by: Christian Stewart <christian@paral.in>

---

v1 -> v2:

 - fix indentation spacing: use tab instead of spaces
 - simplify by using $(1) instead of two separate variables

Signed-off-by: Christian Stewart <christian@paral.in>