package/docker-engine: bump version to v23.0.0
This update requires a workaround for the non-standard vendor/ structure in the
docker-engine repository, which has an invalid vendor/modules.txt file.

Removing the vendor/modules.txt file fixes the build error.

Reference: https://github.com/moby/moby/issues/44618#issuecomment-1343565705

The go download post-process script uses go mod init to create the go.mod file
which specifies a language version of go1.16 on default. Use go mod edit to
adjust the go.mod file to the required minimum version go1.19 for generic types
used by docker-engine. (Fixes a build error). Note that this changes the go.mod
language version and is not required to be equal to the current Go tool version.

Note that as of v23.0.0 docker-engine expects containerd to be started by
systemd if --containerd= flag is set in dockerd.service, whereas in previous
versions containerd was started by the docker daemon as a sub-process.

There is no sysv unit for containerd, but this is OK: in the Docker sysv init
file S60dockerd, we do not set the --containerd flag. The docker daemon will
start containerd as a sub-process automatically due to the absence of this flag.

Signed-off-by: Christian Stewart <christian@paral.in>

---

v1 -> v2:

 - clarify commit message about containerd.service and sysv compat
 - fix formatting in S60dockerd
 - bundle commit into a series with required containerd changes