boot/refind-bin: new package
rEFInd is a user-friendly EFI GUI for boot selection.

It can be used as an alternative to Grub with a user-friendly graphical and/or
text interface for selecting a boot option.

The default refind.conf interactively displays a list of boot options on
startup, including detected Buildroot Linux kernel/initrd files.

https://www.rodsbooks.com/refind/

Signed-off-by: Christian Stewart <christian@aperture.us>