package/docker-cli: bump version to v23.0.0
docker-cli does not ship a go.mod file in the tarball: create one to tell the Go
compiler the root import path prefix for the module and the language version.
Note that the language version in go.mod is not required to be equal the version
of the Go compiler.

This update requires a workaround for the non-standard vendor/ structure in the
docker-engine repository, which has an invalid vendor/modules.txt file.

Reference: https://github.com/moby/moby/issues/44618#issuecomment-1343565705

Signed-off-by: Christian Stewart <christian@paral.in>

---

v2 -> v3:

 - create go.mod file in the CONFIGURE step
 - update comments regarding adjusting language version
 - use rm -f instead of checking for file existence