package/docker-cli-buildx: new package
buildx is a docker CLI plugin implementing extended build features.

buildx is the default builder as of Docker v23.x, and if the plugin is not
present, the CLI will output an error warning that it is using the deprecated
built-in builder instead of the buildx builder.

However, buildx is a separate repository from docker-cli: it is a plugin that
needs to be installed at /usr/lib/docker/cli-plugins on the target.

https://github.com/docker/buildx

Signed-off-by: Christian Stewart <christian@aperture.us>

---

v1 -> v2:

 - buildx: bump version to v0.11.1
 - https://github.com/docker/buildx/releases/tag/v0.11.1
 - NOTE: docker-cli-buildx requires language features from go1.20!
 - Depends on bump to go1.20.x:
   - https://patchwork.ozlabs.org/project/buildroot/patch/20230711220849.1702358-3-christian@aperture.us/

v2 -> v3:

 - Rename bin from buildx to docker-buildx (fixes docker cli not finding it)
 - update docker-cli-buildx to v0.11.2
   - https://github.com/docker/buildx/releases/tag/v0.11.2

v3 -> v4:

 - update docker-cli-buildx to v0.14.0

Signed-off-by: Christian Stewart <christian@aperture.us>