configs/stm32f429_disco_xip: add hashes
The patch adds the .hash file for Linux, which then allow to enable
BR2_DOWNLOAD_FORCE_CHECK_HASHES. With this, we can now drop the
defconfig from .checkpackageignore.

Note that the defconfig uses linux-headers that are not the same as
the kernel (although the same series), and instead use the 6.1.x
version from the linux-headers package choice. So we don't need to
add a .hash file for linux-headers.

Signed-off-by: Dario Binacchi <dario.binacchi@amarulasolutions.com>
[yann.morin.1998@free.fr: don't add linux-headers symlink; explain why]
Signed-off-by: Yann E. MORIN <yann.morin.1998@free.fr>