package/go: bump to version 1.24.4
The latest Go release, version 1.24, arrives in February 2025, six months after
Go 1.23. Most of its changes are in the implementation of the toolchain,
runtime, and libraries. The release maintains the Go 1 promise of compatibility.

go1.24.2 (released 2025-04-01) includes security fixes to the net/http package,
as well as bug fixes to the compiler, the runtime, the go command, and the
crypto/tls, go/types, net/http, and testing packages.

https://go.dev/doc/devel/release#go1.24.2
https://github.com/golang/go/issues?q=milestone%3AGo1.24.2+label%3ACherryPickApproved

go1.24.3 (released 2025-05-06) includes security fixes to the os package, as
well as bug fixes to the runtime, the compiler, the linker, the go command,
and the crypto/tls and os packages.

https://go.dev/doc/devel/release#go1.24.3
https://github.com/golang/go/issues?q=milestone%3AGo1.24.3+label%3ACherryPickApproved

go1.24.4 (released 2025-06-05) includes security fixes to the crypto/x509, net/http, and os packages, as well as bug fixes to the linker, the go command, and the hash/maphash and os packages.

https://go.dev/doc/devel/release#go1.24.4
https://github.com/golang/go/issues?q=milestone%3AGo1.24.4+label%3ACherryPickApproved

The following CVEs are addressed:

CVE-2025-4673: net/http: sensitive headers not cleared on cross-origin redirect
CVE-2025-0913: os: inconsistent handling of O_CREATE|O_EXCL on Unix and Windows
CVE-2025-22874: crypto/x509: usage of ExtKeyUsageAny disables policy validation

Signed-off-by: Christian Stewart <christian@aperture.us>