package/go: bump to version 1.24.2
The latest Go release, version 1.24, arrives in February 2025, six months after
Go 1.23. Most of its changes are in the implementation of the toolchain,
runtime, and libraries. The release maintains the Go 1 promise of compatibility.

go1.24.2 (released 2025-04-01) includes security fixes to the net/http package,
as well as bug fixes to the compiler, the runtime, the go command, and the
crypto/tls, go/types, net/http, and testing packages.

https://go.dev/doc/devel/release#go1.24.2
https://github.com/golang/go/issues?q=milestone%3AGo1.24.2+label%3ACherryPickApproved

Signed-off-by: Christian Stewart <christian@aperture.us>

===

v1 => v2:

 - fix license hash
 - fix dependencies on STAGE3