package/go: bump to version 1.25.1
Building Go 1.24 and later requires Go 1.22.6 or later for bootstrap.

To support this we use our previous Go version 1.23.12 as the version
for go-bootstrap-stage4 and have the build for Go 1.25.1 depend on
go-bootstrap-stage4.

Go version 1.23.12 is the latest go version we can build using
go-bootstrap-stage3.

The package build for go-bootstrap-stage4 if effectively identical to
go-bootstrap-stage3 with only the Go version and stage number changed.

Signed-off-by: James Hilliard <james.hilliard1@gmail.com>